Fix casting tests to use long long instead of long, since, apparently, sizeof(long) == sizeof(int), so the tests did not exercise extension as they should have.

+++ otter/test/TestOtterIntegration/OtterCore/Fast/Casts/bool_to_long_long_to_short.c
@@ -1,11 +1,11 @@
-/* This tests truncation followed by extension. */
+/* This tests extension followed by truncation. */
 
 #pragma no_other_abandoned
 
 int main(void) {
     int x;
-    short y;
-    long z;
+    long long y;
+    short z;
     __SYMBOLIC(&x);
     __ASSUME(x == 0);
     y = !x;
